using cursor instead of curse, now using dynamic imports for extensions
